+++ tunedune.py
@@ -2,17 +2,14 @@ import kivy
 from kivy.app import App
 from kivy.lang import Builder
 from kivy.uix.screenmanager import ScreenManager, Screen
+from loginscreen import LoginScreen
+from mainscreen import MainScreen
 
 kivy.require('1.10.1')
 
 Builder.load_file('sidebar.kv')
-
-class LoginScreen(Screen):
-    pass
-
-class MainScreen(Screen):
-    pass
-
+Builder.load_file('loginscreen.kv')
+Builder.load_file('mainscreen.kv')
 
 class TuneDuneApp(App):
     def build(self):